Improve the command utility
This utility was overly complex for no reason. There was no need for
command objects since we didn't need to store any state.

Also, this improves error handling by throwing an error when the
command exit code is invalid. An additional callback was required to
execute a function before this error was thrown.